Pull out the form into seperate component
